[goecharger] Fix compatibility with go-eCharger V3 (#12733)
Also avoid the thing to remain in state UNKNOWN if the decoding of the
JSON response is failing.

Fix #12719

+++ bundles/org.openhab.binding.goecharger/src/main/java/org/openhab/binding/goecharger/internal/api/GoEStatusResponseV2DTO.java
@@ -34,7 +34,7 @@ public class GoEStatusResponseV2DTO extends GoEStatusResponseBaseDTO {
     public Double[] temperatures;
 
     @SerializedName("wh")
-    public Long sessionChargeConsumption;
+    public Double sessionChargeConsumption;
 
     @SerializedName("dwo")
     public Double sessionChargeConsumptionLimit;
